bpo-XXX: Allow deferred flushing
MemoryHandler try to buffer messages and handle them to a target handler
to minimize logging overhead. However if the target handler flushes each
record in handle(), we don't gain much by buffering the messages.

Add "autoflush" class attribute to Handler class, and use it to avoid
automatic flushing in StreamHandler.emit(). This will buffer the
formatted messages in the underlying stream, and flush them to storage
only when the buffer is full, or the user call flush explicitly.
MemoryHandler uses the new attribute to flush once after handling all
pending messages to the target handler.

This patch does not change the behavior of existing code. To use the new
feature, you must create a handler subclass overriding "autoflush" to
False.